legacy/pyproject.toml
[project]
name = "vbn"
version = "0.1.0"
description = "Local-first analysis tools for Allen Institute Visual Behavior Neuropixels dataset"
requires-python = ">=3.10"
dependencies = [
    "allensdk>=2.16.0",
    "pandas==1.5.3",          # ✅ Match AllenSDK requirement exactly
    "numpy<1.24",             # Compatible with pandas 1.5.3
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",      # Compatible with pandas 1.5.3
    "h5py>=3.9.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
sleap = ["sleap>=1.3.0"]
dlc = ["deeplabcut>=2.3.0"]
pose = ["vbn[sleap]"]
dev = ["pytest", "ipykernel", "jupyterlab", "pandas==1.5.3"]  # Ensure consistency

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"